Updated procfs to work on 10.5 (will no longer compile on 10.4, actually).

+++ filesystems/procfs/procfs_displays.cc
@@ -31,7 +31,7 @@ CreateScreenGLContext(CGDirectDisplayID targetDisplay)
 {
     CGOpenGLDisplayMask displayMask;
     CGLPixelFormatObj   pixelFormatObj;
-    long                numPixelFormats;
+    GLint               numPixelFormats;
 
     displayMask = CGDisplayIDToOpenGLDisplayMask(targetDisplay);
